selecting a time limit
